Move practically everything into `src/ethers/` submodule
