+++ src/viem/index.ts
@@ -1,0 +1,1 @@
+export const placeholder = null;
